feat: Add SolarCoin and LunarToken contracts
Add SolarCoin.sol and LunarToken.sol contracts to the project. These contracts are ERC20 tokens that inherit from the OpenZeppelin ERC20 contract. They have a constructor that mints an initial supply of tokens to the contract deployer and a mint function that can be called by the contract owner to mint additional tokens.

Also, update the package.json and package-lock.json files to include the "@openzeppelin/contracts" dependency.